+++ GetThePicture/Picture/Clause/Base/Computational/COMP5-NativeBinary.cs
@@ -1,14 +1,17 @@
-using GetThePicture.Picture.Clause.Base.Options;
 using GetThePicture.Picture.Clause.Codec.Category.Numeric;
 
 namespace GetThePicture.Picture.Clause.Base.Computational;
 
 /// <summary>
-/// COMP-5 (Native Binary)
+/// COMP-5 (Native Binary) <br/>
+/// - Uses platform-native endian. <br/>
+/// - Little Endian on x86/x64, Big Endian on mainframe. <br/>
+/// - Cross-platform use: handle endian conversion as needed. <br/>
+/// - For consistency with COMP-4/COMP-3, codec may optionally normalize to Big Endian. <br/>
 /// </summary>
 internal static class COMP5
 {
-    public static object Decode(ReadOnlySpan<byte> buffer, PicMeta pic, BinaryOptions endian = BinaryOptions.Normal)
+    public static object Decode(ReadOnlySpan<byte> buffer, PicMeta pic, bool isBigEndian = true)
     {
         if (pic.DecimalDigits > 0)
             throw new NotSupportedException($"COMP-5 does not support decimal digits. PIC has {pic.DecimalDigits} decimal digits.");
@@ -21,7 +24,7 @@ internal static class COMP5
         Span<byte> bytes = stackalloc byte[length];
         buffer[..length].CopyTo(bytes);
 
-        if (endian == BinaryOptions.Reversed)
+        if (BitConverter.IsLittleEndian && isBigEndian)
             bytes.Reverse();
 
         return length switch
@@ -37,7 +40,7 @@ internal static class COMP5
         };
     }
 
-    public static byte[] Encode(NumericMeta nMeta, PicMeta pic, BinaryOptions endian = BinaryOptions.Normal)
+    public static byte[] Encode(NumericMeta nMeta, PicMeta pic, bool isBigEndian = true)
     {
         if (pic.DecimalDigits > 0)
             throw new NotSupportedException($"COMP-5 does not support decimal digits. PIC has {pic.DecimalDigits} decimal digits.");
@@ -90,7 +93,7 @@ internal static class COMP5
             _ => throw new NotSupportedException()
         };
 
-        if (endian == BinaryOptions.Reversed)
+        if (BitConverter.IsLittleEndian && isBigEndian)
             bytes.Reverse();
 
         return bytes.ToArray();
@@ -103,7 +106,7 @@ internal static class COMP5
             <=  4 => 2,
             <=  9 => 4,
             <= 18 => 8,
-            _ => throw new NotSupportedException("Too many digits for COMP-5 (Native-Binary)")
+            _ => throw new NotSupportedException("Too many digits for COMP-4 (Binary) or COMP-5 (Native-Binary)")
         };
     }
 }